Split src/commands/run.rs into submodule directory
Break the 1,417-line run.rs into 9 focused files under src/commands/run/:
- mod.rs: orchestrator (handle_run setup + delegation)
- constants.rs: queue sizes, worker counts, queue_warning_threshold()
- aprs.rs: process_aprs_message()
- beast.rs: process_beast_message() + extract_icao_from_message()
- sbs.rs: process_sbs_message()
- workers.rs: spawn functions for OGN, Beast, SBS, and APRS processor workers
- monitoring.rs: spawn_metrics_reporter() for queue depth + DB pool metrics
- shutdown.rs: spawn_shutdown_handler() for graceful Ctrl+C drain
- tests.rs: timestamp parsing tests

Pure move refactor — no logic changes. Follows the existing pattern
established by commands/load_data/ and commands/archive/.
